Update players list and standings every time scores are recalculated
