+++ src/Shared/Actions.js
@@ -18,6 +18,7 @@ const ROUNDS_DELETE_LAST            = 'Rounds/DeleteLast'
 const ROUNDS_CHANGE_MATCH_RESULT    = 'Rounds/ChangeMatchResult'
 const ROUNDS_UPDATE_CUSTOM_MATCHES  = 'Rounds/UpdateCustomMatches'
 const STANDINGS_CHANGE_STAT         = 'Standings/ChangeStat'
+const STANDINGS_REFRESH             = 'Standings/Refresh'
 const TOURNAMENT_START_NEW          = 'Tournament/StartNew'
 const TOURNAMENT_HIDE_LOAD_MODAL    = 'Tournament/HideLoadModal'
 const TOURNAMENT_HIDE_SAVE_MODAL    = 'Tournament/HideSaveModal'
@@ -55,6 +56,7 @@ export const Types = {
     ROUNDS_CHANGE_MATCH_RESULT,
     ROUNDS_UPDATE_CUSTOM_MATCHES,
     STANDINGS_CHANGE_STAT,
+    STANDINGS_REFRESH,
     TOURNAMENT_START_NEW,
     TOURNAMENT_HIDE_LOAD_MODAL,
     TOURNAMENT_HIDE_SAVE_MODAL,
