+++ src/Standings/ActionCreators/RefreshStandings.js
@@ -1,0 +1,4 @@
+import { create, Types } from '../../Shared/Actions'
+
+export default () => create(Types.STANDINGS_REFRESH)()
+
